refactor: extract notification scheduling
